backend_python/requirements.txt
fastapi==0.115.12
pydantic-settings==2.9.1
uvicorn==0.34.2